WIP: Avoid build failure on OSX 10.4.
The Availability.h header didn't appear until 10.5.  The macro
__ENVIRONMENT_MAC_OS_X_VERSION_MIN_REQUIRED__ also didn't appear (by
default) until 10.5, but when present it's defined directly by the
compiler, so it's testable without the include.  In all cases, it's
defined when the compiler is run with the 'macosx-version-min' option,
and when not defined, the default value of 0 has the intended effect.
Hence, using this macro avoids the need for the troublesome header,
and there's no need for an explicit defined() check.

This fix applies the same treatment to iOS, though that's untested.

***TESTED (Pending):
Built on macOS 10.4-10.5 ppc, 10.4-10.6 i386, and 10.6-10.15 x86_64.
Verified that libcrypto references _CCRandomGenerateBytes in macOS
versions >=10.12 and not otherwise.